New Plugin: Apollo-Android (Java) (#1924)
* initial commit
supports input types
changes to the core
added basic preset with support for this

* WIP operations

* WIP fixes

* wip

* fixes

* wip

* fixes after rebase
cleanup

* fix operation marshaller

* fixes for marshaller

* fix syntax errors

* fix nested mappers issues
fixes for list types

* fixes for filenames
added fragments files stub

* fixes

* added missing arguments

* fixes

* fixes

* fixes

* fixes for input visitor

* cleanup

* added support for fragments

* added CustomType class generator

* update lockfile

* added a note regarding client-side plugins

* update lockfile

* fixes after rebase

* v1.2.2-beta.0

* minor fix

* update lockfile

* fixes for documents validation issues

* fixes for output path

* fix output path for type

* replace float with double

* try to fix issues with output

* fixes for output dir

* fix deps

* fixes for deps
fixes for usage

* fixes

* fixes

* some fixes for generated types

* fixes for tests

* fix object list type

* fix issues with lists

* fixes
added tests

* fixes for issues with fragments

* test

* fixes

* rebase and fixes

* fix issues with fragments and readConditional

* updated docs

+++ packages/plugins/java/common/src/scalars.ts
@@ -3,5 +3,5 @@ export const JAVA_SCALARS = {
   String: 'String',
   Boolean: 'Boolean',
   Int: 'Integer',
-  Float: 'Float',
+  Float: 'Double',
 };
